% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_merge_mismatch.R
\name{detect_merge_mismatch}
\alias{detect_merge_mismatch}
\title{Detect Merge Mismatches}
\usage{
detect_merge_mismatch(var1, var2)
}
\arguments{
\item{var1}{a variable from the larger data set to be merged}

\item{var2}{a variable from the smaller data set to be merged}
}
\value{
a data.frame containing var1 and grep matches of var1 in var2
}
\description{
Detect Merge Mismatches
}
\examples{
detect_merge_mismatch(c("species", "monkey", "human"),
                      c("human bowling ball", "rabid monkey", "devoloution"))

}

