% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_intpen.R
\name{ggplot_intpen}
\alias{ggplot_intpen}
\title{Plot Internet Penetration Data by Country and Year}
\usage{
ggplot_intpen(year = 2014, data1 = WorldMapInternetPenetration,
  gradient.low = "white", gradient.high = "#649AF1")
}
\arguments{
\item{year}{the year to plot}

\item{data1}{use the default database, WorldMapInternetPenetration, for
Internet penetration data}

\item{gradient.low}{the lowest color value for country and legend fill}

\item{gradient.high}{the highest color value for country and legend fill}
}
\value{
a ggplot object
}
\description{
Plot data from the World Bank's World Development Indicators on a
geom_polygon representation of the countries of the world.
}
\examples{
ggplot_intpen(year = 2014)
}

