% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IP_country.R
\name{IP_code}
\alias{IP_code}
\title{Convert IP addresses to IP codes}
\usage{
IP_code(IP.address)
}
\arguments{
\item{IP.address}{a character or factor vector of one or more IP addresses}
}
\value{
Returns a numeric vector of IP codes corresponding to
  \code{IPaddress}
}
\description{
Convert IP addresses to IP codes
}
\examples{
IP_code(ips)

}

