% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IP_country.R
\name{IP_country}
\alias{IP_country}
\title{Convert IP addresses to country names}
\usage{
IP_country(IP.address, IP.database = NULL)
}
\arguments{
\item{IP.address}{a character or factor vector of one or more IP addresses}

\item{IP.database}{an IP database, see ?IP.database}
}
\value{
Returns a factor vector of country names corresponding to
  \code{IPaddress}
}
\description{
Convert IP addresses to country names
}
\examples{
mydata <- IP_generator(1000)
IP_country(mydata)


}

