% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IP_country.R
\name{IP_integer}
\alias{IP_integer}
\title{Convert IP address to IP integer}
\usage{
IP_integer(IP.address)
}
\arguments{
\item{IP.address}{a character or factor vector of one or more IP addresses}
}
\value{
Returns a numeric vector of IP integers calculated from \code{IPaddress}
  by spliting the string
}
\description{
Convert IP address to IP integer
}
\examples{
IP_integer(IPs)

}

