% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IP_country.R
\name{IP_lookup}
\alias{IP_lookup}
\title{Match IP codes in a IP database}
\usage{
IP_lookup(IP.code, IP.database = NULL)
}
\arguments{
\item{IP.code}{a character or factor vector of one or more IP addresses}

\item{IP.database}{an IP database, see ?IP.database}
}
\value{
Returns country from IP codes
}
\description{
Match IP codes in a IP database
}
\examples{
ips.code <- IP_code(ips)
IP_lookup(ips.code)
}

