% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotIPs.R
\name{plotIPs}
\alias{plotIPs}
\title{Title
Convert IPs to countries and plot on world map}
\usage{
plotIPs(sample_IPs)
}
\arguments{
\item{sample_IPs}{a vector of IP addresses}
}
\value{
Returns a gradient map, indicating % of IP addresses originating from each country
}
\description{
Title
Convert IPs to countries and plot on world map
}
\examples{
plotIPs(sample_IPs)
}

