% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IP_country.R
\name{IP_country}
\alias{IP_country}
\title{Convert IP addresses to country names}
\usage{
IP_country(IPaddress)
}
\arguments{
\item{IPaddress}{A vector of IP addresses}
}
\value{
Returns a vector of country names for the locations of \code{IPaddress}
}
\description{
Convert IP addresses to country names
}
\examples{
IP_country(IP_generator(1000))


}

