% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IP_country.R
\name{IP_lookup}
\alias{IP_lookup}
\title{Title}
\usage{
IP_lookup(x, y)
}
\arguments{
\item{x}{IP database}

\item{y}{IP addresses}
}
\value{
Returns country from IP codes
}
\description{
Title
}
\examples{
data1 <- IP_code(ips)
IP_lookup(data1)
}

